eliminar y editar bibliotecas

+++ backend/modelos/modelos_bibliotecas.js
@@ -41,7 +41,6 @@ async function agregar_libro_biblioteca(username_cliente, id_biblioteca, isbn_co
     }
     try{    
         const libro_agregado = await db_client.query('INSERT INTO biblioteca_libro (id_biblioteca, isbn_code) VALUES ($1, $2) RETURNING isbn_code',[id_biblioteca, isbn_code]);
-        console.log(`el libro es ${libro_agregado}`);
         return libro_agregado.rows[0];
     }
     catch(error){
@@ -57,9 +56,12 @@ async function agregar_libro_biblioteca(username_cliente, id_biblioteca, isbn_co
     }
 }
 
-async function eliminar_biblioteca(id_biblioteca) {
+async function eliminar_biblioteca(username_cliente, id_biblioteca) {
+    const username_credor = await get_creador_biblioteca(id_biblioteca);
+    if (username_credor === undefined || username_credor !== username_cliente){
+        return {status: 401, error: "No eres dueño de esta biblioteca"};
+    }
 
-    console.log(id_biblioteca);
     const biblioteca_eliminada = await db_client.query('DELETE FROM bibliotecas WHERE id_biblioteca = $1 RETURNING nombre_biblioteca',[id_biblioteca]);
     if (biblioteca_eliminada.rowCount === 0){
         return undefined;
@@ -67,7 +69,11 @@ async function eliminar_biblioteca(id_biblioteca) {
     return biblioteca_eliminada.rows[0].nombre_biblioteca; 
 }
 
-async function eliminar_libro_biblioteca(id_biblioteca, isbn_code) {
+async function eliminar_libro_biblioteca(username_cliente, id_biblioteca, isbn_code) {
+    const username_credor = await get_creador_biblioteca(id_biblioteca);
+    if (username_credor === undefined || username_credor !== username_cliente){
+        return {status: 401, error: "No eres dueño de esta biblioteca"};
+    }
 
     const libro_eliminado = await db_client.query('DELETE FROM biblioteca_libro WHERE isbn_code = $1 AND id_biblioteca = $2 RETURNING isbn_code',[isbn_code, id_biblioteca]);
     if (libro_eliminado.rowCount === 0){
@@ -76,12 +82,17 @@ async function eliminar_libro_biblioteca(id_biblioteca, isbn_code) {
     return libro_eliminado.rows[0].isbn_code; 
 }
 
-async function modificar_biblioteca(id_biblioteca, nombre_biblioteca, icono = null) {
-    const libro_modificado = await db_client.query('UPDATE bibliotecas SET nombre_biblioteca = $1, icono = $2 WHERE id_biblioteca = $3 RETURNING nombre_biblioteca',[nombre_biblioteca, icono, id_biblioteca]);
-    if (libro_modificado.rowCount === 0){
+async function modificar_biblioteca(username_cliente, id_biblioteca, nombre_biblioteca, icono = null) {
+    const username_credor = await get_creador_biblioteca(id_biblioteca);
+    if (username_credor === undefined || username_credor !== username_cliente){
+        return {status: 401, error: "No eres dueño de esta biblioteca"};
+    }
+
+    const biblioteca_actualizada = await db_client.query('UPDATE bibliotecas SET nombre_biblioteca = $1, icono = $2 WHERE id_biblioteca = $3 RETURNING nombre_biblioteca',[nombre_biblioteca, icono, id_biblioteca]);
+    if (biblioteca_actualizada.rowCount === 0){
         return undefined;
     }
-    return libro_modificado.rows[0].nombre_biblioteca;
+    return biblioteca_actualizada.rows[0].nombre_biblioteca;
 }
